requirements.txt
coverage==4.5.2
Django==2.1.3
django-cors-middleware==1.3.1
django-extensions==2.1.4
django-rest-swagger==2.1.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
django-taggit==0.23.0
gunicorn==19.9.0
social-auth-app-django==3.1.0
psycopg2-binary==2.7.6.1
PyJWT==1.6.4
pytz==2018.7
whitenoise==4.1.2
six==1.11.0
